transform Segments to visible JPanel

+++ src/GUISegment.java
@@ -1,4 +1,4 @@
-import javax.swing.*;
+/*import javax.swing.*;
 import java.awt.*;
 
 public class GUISegment extends JPanel {
@@ -8,12 +8,13 @@ public class GUISegment extends JPanel {
     private int y2;
 
     GUISegment(Segment s) {
-        this.x1 = (int) Math.round(s.getP0().mX);
-        this.y1 = (int) Math.round(s.getP0().mY);
-        this.x2 = (int) Math.round(s.getP1().mX);
-        this.y2 = (int) Math.round(s.getP1().mY);
+        this.x1 = Math.round(s.getP0().mX);
+        this.y1 = Math.round(s.getP0().mY);
+        this.x2 = Math.round(s.getP1().mX);
+        this.y2 = Math.round(s.getP1().mY);
     }
     public void paint(Graphics2D g2D) {
         g2D.drawLine(this.x1, this.y1, this.x2, this.y2);
     }
 }
+*/
